Chore: speed up local dev build
